Relocated component and exported for public use.
